moved image management to diffrent file
